started on user editing interface

+++ app/Http/Controllers/UserController.php
@@ -26,7 +26,7 @@ class UserController extends Controller
 
     public function getUpdate($id) {
         $currentUser = Auth::user();
-        // Restrict non-admin access. This should be a middleware for security reasons, but that's a lot of work and this is area is already protected by Auth...
+        // Restrict non-admin access. This should be a middleware for security reasons, but that's a lot of work and this is area is already protected by Authentication...
         if ((!$currentUser->isAdmin()) && $currentUser->id != $id) {
           return redirect()->route('admin.index');
         }
@@ -35,10 +35,17 @@ class UserController extends Controller
     }
 
     public function postUpdate(Request $request) {
-        
+        $request->validate([
+            'name' => 'required',
+            
+        ]);
+        return redirect()->route(Auth::user()->isAdmin() ? 'users.index' : 'admin.index')->with('message', 'Account updated!');
     }
 
     public function getDelete($id) {
+        if(Auth::user()->id == $id) {
+            return redirect()->route('users.index')->with('message', 'That is the current user. Can\'t remove.');
+        }
         $user = User::find($id);
         $user->delete();
         return redirect()->route('users.index')->with('message', 'User removed!');
